Testing JSOn Recursion

+++ Lesson29-AzureServiceBus/Example/main.go
@@ -11,6 +11,8 @@ import (
 func main() {
 	log.Printf("Starting")
 
+	var readMessages, writeMessages, deleteMessages = true, false, true
+
 	const (
 		azuresb_ns         = "cnSBTest"
 		azuresb_policyname = "ManageTest"
@@ -24,62 +26,46 @@ func main() {
 	client := asbclient.New(asbclient.Queue, azuresb_ns, azuresb_policyname, azuresb_policyval)
 	path := azuresb_queue
 
-	i := 0
-	for {
+	if writeMessages {
+		i := 0
+		for {
 
-		log.Printf("Send: %d", i)
-		err := client.Send(path, &asbclient.Message{
-			Body: []byte(fmt.Sprintf("message %d", i)),
-		})
+			log.Printf("Send: %d", i)
+			err := client.Send(path, &asbclient.Message{
+				Body: []byte(fmt.Sprintf("message %d", i)),
+			})
 
-		if err != nil {
-			log.Printf("Send error: %s", err)
-		} else {
-			log.Printf("Sent: %d", i)
-		}
-
-		time.Sleep(time.Millisecond * 500)
-		i++
-	}
-
-	// Commented Concurrency code
-	// go func() {
-	// 	i := 0
-	// 	for {
-
-	// 		log.Printf("Send: %d", i)
-	// 		err := client.Send(path, &asbclient.Message{
-	// 			Body: []byte(fmt.Sprintf("message %d", i)),
-	// 		})
-
-	// 		if err != nil {
-	// 			log.Printf("Send error: %s", err)
-	// 		} else {
-	// 			log.Printf("Sent: %d", i)
-	// 		}
-
-	// 		time.Sleep(time.Millisecond * 500)
-	// 		i++
-	// 	}
-	// }()
-
-	for {
-		log.Printf("Peeking...")
-		msg, err := client.PeekLockMessage(path, 30)
-
-		if err != nil {
-			log.Printf("Peek error: %s", err)
-		} else {
-			log.Printf("Peeked message: '%s'", string(msg.Body))
-			err = client.DeleteMessage(msg)
 			if err != nil {
-				log.Printf("Delete error: %s", err)
+				log.Printf("Send error: %s", err)
 			} else {
-				log.Printf("Deleted message")
+				log.Printf("Sent: %d", i)
 			}
-		}
 
-		time.Sleep(time.Millisecond * 200)
+			time.Sleep(time.Millisecond * 500)
+			i++
+		}
 	}
 
+	if readMessages {
+		for {
+			log.Printf("Peeking...")
+			msg, err := client.PeekLockMessage(path, 30)
+
+			if err != nil {
+				log.Printf("Peek error: %s", err)
+			} else {
+				log.Printf("Peeked message: '%s'", string(msg.Body))
+				if deleteMessages {
+					err = client.DeleteMessage(msg)
+					if err != nil {
+						log.Printf("Delete error: %s", err)
+					} else {
+						log.Printf("Deleted message")
+					}
+				}
+			}
+
+			//time.Sleep(time.Millisecond * 200)
+		}
+	}
 }
